filesystem/requirements.txt
Flask
Werkzeug
python-dotenv
requests
PyJWT